// frontend/src/api/generated/model/maintenanceCommentInput.ts
export interface MaintenanceCommentInput {
  /** @maxLength 1000 */
  comment?: string;
}